backendff/requirements.txt
fastapi==0.75.1
uvicorn==0.17.6

pydantic~=1.9.0

kafka-python==2.0.2